fix(protondb,gid,ui): live-API score floats, huge egstore manifests, B1 trap; dead code
Real-library probe fallout: the live ProtonDB API returns score as a
float (0.82) which broke every summary decode; real .egstore markers
carry multi-MB FileManifestLists so ParseEGStoreManifest now streams
the header and stops early; fuzzy skips candidates under 4 normalized
chars ('b1' exact-matching the unrelated store item 'B1'); the Epic
catalog id now reaches rows from scanLevel too (it only flowed through
the root-row path); and the unused gameTitle alias plus two alias
doc-comment forms are cleaned up for staticcheck.

+++ internal/protondb/client.go
@@ -50,15 +50,16 @@ const (
 	cooldownFile = "cooldown.json"
 )
 
-// Summary mirrors the ProtonDB summary JSON fields actually used.
+// Summary mirrors the ProtonDB summary JSON fields actually used. Score
+// is a float on the live API (0.82), not an int.
 type Summary struct {
-	Tier             string `json:"tier"`
-	Confidence       string `json:"confidence"`
-	Score            int    `json:"score"`
-	Total            int    `json:"total"`
-	BestReportedTier string `json:"bestReportedTier"`
-	TrendingTier     string `json:"trendingTier"`
-	ProvisionalTier  string `json:"provisionalTier,omitempty"`
+	Tier             string  `json:"tier"`
+	Confidence       string  `json:"confidence"`
+	Score            float64 `json:"score"`
+	Total            int     `json:"total"`
+	BestReportedTier string  `json:"bestReportedTier"`
+	TrendingTier     string  `json:"trendingTier"`
+	ProvisionalTier  string  `json:"provisionalTier,omitempty"`
 }
 
 // Client fetches ProtonDB report summaries.
